05-ComponentScan-example/src/main/java/com/kejeiri/courses/didemo/app/controllers/SetterInjectedController.java
package com.kejeiri.courses.didemo.app.controllers;

import com.kejeiri.courses.services.GreetingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Controller;

@Controller
public class SetterInjectedController {

    private GreetingService greetingService;

    public String sayHello(){
        return greetingService.sayGreeting();
    }

    //since we have three beans (services implementing the Greeting interface) we need to tell spring
    // which bean to use by using the @Qualifier annotation and use the camel case service name
    @Autowired
    public void setGreetingService(@Qualifier("setterGreetingService") GreetingService greetingService) {
        this.greetingService = greetingService;
    }

}
